Camera: Add a GPS indicator
This will add a GPS indicator to the preview of the photo camera.
The indicator is placed as an overlay at the upper top corner
of the photo camera preview.

It shows one of three icons depending on the GPS state:
    - acquiring fix
    - GPS fix acquired
    - GPS disallowed in system settings

The icon will only be shown in the preview of the photo module,
not in the video or panorama module, and not in the gallery view.

Drawables by blunden.

Change-Id: I0686ef4ee1fe9b34ac9beaaa2bd43555234f4f4d